Change type on params

+++ src/main/kotlin/firebase/admin/database/Reference.kt
@@ -30,8 +30,8 @@ external interface Reference {
     fun remove(): Promise<Any?>
     fun set(value: Any): Promise<Any?>
     fun setWithPriority(value: Any, newPriority: Any): Promise<Any?>
-    fun startAt(index: Int): Query
-    fun startAt(index: Int, key: String): Query
+    fun startAt(index: Double): Query
+    fun startAt(index: Double, key: String): Query
     fun startAt(value: String): Query
     fun startAt(value: Boolean): Query
     fun startAt(value: Boolean, key: String): Query
